requirements/base.txt
pytz==2022.7  # https://github.com/stub42/pytz
Pillow==9.3.0  # https://github.com/python-pillow/Pillow        
whitenoise==6.2.0  # https://github.com/evansd/whitenoise

# Django
# ------------------------------------------------------------------------------
django==4.2.16  # pyup: < 4.1  # https://www.djangoproject.com/
django-model-utils==4.3.1  # https://github.com/jazzband/django-model-utils
django-crispy-forms==1.14.0  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==0.7  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-filter==23.5 # https://github.com/carltongibson/django-filter

# Django REST Framework
djangorestframework==3.14.0  # https://github.com/encode/django-rest-framework
django-cors-headers==3.13.0 # https://github.com/adamchainz/django-cors-headers
django-modeltranslation==0.17.0 
# PDF generator
reportlab==4.0.4
xhtml2pdf==0.2.15

# Customize django admin
django-jet-reboot==1.3.5

# Environment variable
python-decouple==3.8

# Celery/redis
celery==5.4.0
redis==5.0.4
hiredis==2.3.2

# Payments
stripe==5.5.0
gopay==1.3.1
